requirements.txt
aiogram~=3.5.0
python-dotenv~=1.0.1
pip~=24.0
psycopg2-binary~=2.9.9